Parse out the reusable logic from infer_headers and variant_to_bigquery modules into libs directory.
