+++ gcp_variant_transforms/libs/processed_variant.py
@@ -38,7 +38,6 @@ from gcp_variant_transforms.libs import bigquery_util
 from gcp_variant_transforms.libs import bigquery_sanitizer
 from gcp_variant_transforms.libs.annotation import annotation_parser
 from gcp_variant_transforms.libs.annotation.vep import descriptions
-from gcp_variant_transforms.transforms import infer_headers
 
 _FIELD_COUNT_ALTERNATE_ALLELE = 'A'
 
@@ -354,7 +353,7 @@ class ProcessedVariantFactory(object):
     annotation_names = annotation_parser.extract_annotation_names(
         self._header_fields.infos[annot_field][_HeaderKeyConstants.DESC])
     for name in annotation_names:
-      type_key = infer_headers.get_inferred_annotation_type_header_key(
+      type_key = annotation_parser.get_inferred_annotation_type_header_key(
           annot_field, name)
       yield name, type_key
 
@@ -476,7 +475,7 @@ class _AnnotationProcessor(object):
         for name, value in annotation_map.iteritems():
           if name == annotation_parser.ANNOTATION_ALT:
             continue
-          type_key = infer_headers.get_inferred_annotation_type_header_key(
+          type_key = annotation_parser.get_inferred_annotation_type_header_key(
               annotation_field_name, name)
           vcf_type = self._vcf_type_from_annotation_header(
               annotation_field_name, type_key)
