+++ gcp_variant_transforms/transforms/variant_to_bigquery.py
@@ -16,15 +16,10 @@
 
 from __future__ import absolute_import
 
-import exceptions
 import random
-import re
 from typing import Dict, List  # pylint: disable=unused-import
 
 import apache_beam as beam
-from apache_beam.io.gcp.internal.clients import bigquery
-from apitools.base.py import exceptions
-from oauth2client.client import GoogleCredentials
 
 from gcp_variant_transforms.beam_io import vcf_header_io  # pylint: disable=unused-import
 from gcp_variant_transforms.libs import bigquery_schema_descriptor
@@ -132,7 +127,8 @@ class VariantToBigQuery(beam.PTransform):
     self._omit_empty_sample_calls = omit_empty_sample_calls
     self._num_bigquery_write_shards = num_bigquery_write_shards
     if update_schema_on_append:
-      self._update_bigquery_schema_on_append()
+      bigquery_util.update_bigquery_schema_on_append(self._schema.fields,
+                                                     self._output_table)
 
   def expand(self, pcoll):
     bq_rows = pcoll | 'ConvertToBigQueryTableRow' >> beam.ParDo(
@@ -172,83 +168,3 @@ class VariantToBigQuery(beam.PTransform):
                       beam.io.BigQueryDisposition.WRITE_APPEND
                       if self._append
                       else beam.io.BigQueryDisposition.WRITE_TRUNCATE))))
-
-  def _update_bigquery_schema_on_append(self):
-    # type: (bool) -> None
-    # if table does not exist, do not need to update the schema.
-    # TODO (yifangchen): Move the logic into validate().
-    output_table_re_match = re.match(
-        r'^((?P<project>.+):)(?P<dataset>\w+)\.(?P<table>[\w\$]+)$',
-        self._output_table)
-    credentials = GoogleCredentials.get_application_default().create_scoped(
-        ['https://www.googleapis.com/auth/bigquery'])
-    client = bigquery.BigqueryV2(credentials=credentials)
-    try:
-      project_id = output_table_re_match.group('project')
-      dataset_id = output_table_re_match.group('dataset')
-      table_id = output_table_re_match.group('table')
-      existing_table = client.tables.Get(bigquery.BigqueryTablesGetRequest(
-          projectId=project_id,
-          datasetId=dataset_id,
-          tableId=table_id))
-    except exceptions.HttpError:
-      return
-
-    new_schema = bigquery.TableSchema()
-    new_schema.fields = _get_merged_field_schemas(existing_table.schema.fields,
-                                                  self._schema.fields)
-    existing_table.schema = new_schema
-    try:
-      client.tables.Update(bigquery.BigqueryTablesUpdateRequest(
-          projectId=project_id,
-          datasetId=dataset_id,
-          table=existing_table,
-          tableId=table_id))
-    except exceptions.HttpError as e:
-      raise RuntimeError('BigQuery schema update failed: %s' % str(e))
-
-
-def _get_merged_field_schemas(
-    field_schemas_1,  # type: List[bigquery.TableFieldSchema]
-    field_schemas_2  # type: List[bigquery.TableFieldSchema]
-    ):
-  # type: (...) -> List[bigquery.TableFieldSchema]
-  """Merges the `field_schemas_1` and `field_schemas_2`.
-
-  Args:
-    field_schemas_1: A list of `TableFieldSchema`.
-    field_schemas_2: A list of `TableFieldSchema`.
-  Returns:
-    A new schema with new fields from `field_schemas_2` appended to
-    `field_schemas_1`.
-  Raises:
-    ValueError: If there are fields with the same name, but different modes or
-    different types.
-  """
-  existing_fields = {}  # type: Dict[str, bigquery.TableFieldSchema]
-  merged_field_schemas = []  # type: List[bigquery.TableFieldSchema]
-  for field_schema in field_schemas_1:
-    existing_fields.update({field_schema.name: field_schema})
-    merged_field_schemas.append(field_schema)
-
-  for field_schema in field_schemas_2:
-    if field_schema.name not in existing_fields.keys():
-      merged_field_schemas.append(field_schema)
-    else:
-      existing_field_schema = existing_fields.get(field_schema.name)
-      if field_schema.mode != existing_field_schema.mode:
-        raise ValueError(
-            'The mode of field {} is not compatible. The original mode is {}, '
-            'and the new mode is {}.'.format(field_schema.name,
-                                             existing_field_schema.mode,
-                                             field_schema.mode))
-      if field_schema.type != existing_field_schema.type:
-        raise ValueError(
-            'The type of field {} is not compatible. The original type is {}, '
-            'and the new type is {}.'.format(field_schema.name,
-                                             existing_field_schema.type,
-                                             field_schema.type))
-      if field_schema.type == bigquery_util.TableFieldConstants.TYPE_RECORD:
-        existing_field_schema.fields = _get_merged_field_schemas(
-            existing_field_schema.fields, field_schema.fields)
-  return merged_field_schemas
