+++ gcp_variant_transforms/transforms/variant_to_bigquery_test.py
@@ -33,7 +33,6 @@ from gcp_variant_transforms.libs import vcf_field_conflict_resolver
 from gcp_variant_transforms.libs.bigquery_util import ColumnKeyConstants
 from gcp_variant_transforms.libs.bigquery_util import TableFieldConstants
 from gcp_variant_transforms.testing import vcf_header_util
-from gcp_variant_transforms.transforms import variant_to_bigquery
 from gcp_variant_transforms.transforms.variant_to_bigquery import ConvertVariantToRow
 
 
@@ -277,325 +276,3 @@ class ConvertToBigQueryTableRowTest(unittest.TestCase):
             self._row_generator, allow_incompatible_records=True)))
     assert_that(bigquery_rows, equal_to([row]))
     pipeline.run()
-
-  def test_merge_field_schemas_no_same_id(self):
-    field_schemas_1 = [
-        bigquery.TableFieldSchema(
-            name='II',
-            type=TableFieldConstants.TYPE_INTEGER,
-            mode=TableFieldConstants.MODE_NULLABLE,
-            description='INFO foo desc'),
-        bigquery.TableFieldSchema(
-            name='IFR',
-            type=TableFieldConstants.TYPE_FLOAT,
-            mode=TableFieldConstants.MODE_REPEATED,
-            description='INFO foo desc')
-    ]
-    field_schemas_2 = [
-        bigquery.TableFieldSchema(
-            name='AB',
-            type=TableFieldConstants.TYPE_FLOAT,
-            mode=TableFieldConstants.MODE_NULLABLE,
-            description='INFO foo desc')
-    ]
-    merged_field_schemas = variant_to_bigquery._get_merged_field_schemas(
-        field_schemas_1, field_schemas_2)
-    expected_merged_field_schemas = [
-        bigquery.TableFieldSchema(
-            name='II',
-            type=TableFieldConstants.TYPE_INTEGER,
-            mode=TableFieldConstants.MODE_NULLABLE,
-            description='INFO foo desc'),
-        bigquery.TableFieldSchema(
-            name='IFR',
-            type=TableFieldConstants.TYPE_FLOAT,
-            mode=TableFieldConstants.MODE_REPEATED,
-            description='INFO foo desc'),
-        bigquery.TableFieldSchema(
-            name='AB',
-            type=TableFieldConstants.TYPE_FLOAT,
-            mode=TableFieldConstants.MODE_NULLABLE,
-            description='INFO foo desc')
-    ]
-    self.assertEqual(merged_field_schemas, expected_merged_field_schemas)
-
-  def test_merge_field_schemas_same_id_no_conflicts(self):
-    field_schemas_1 = [
-        bigquery.TableFieldSchema(
-            name='II',
-            type=TableFieldConstants.TYPE_INTEGER,
-            mode=TableFieldConstants.MODE_NULLABLE,
-            description='INFO foo desc'),
-        bigquery.TableFieldSchema(
-            name='IFR',
-            type=TableFieldConstants.TYPE_FLOAT,
-            mode=TableFieldConstants.MODE_REPEATED,
-            description='INFO foo desc')
-    ]
-    field_schemas_2 = [
-        bigquery.TableFieldSchema(
-            name='II',
-            type=TableFieldConstants.TYPE_INTEGER,
-            mode=TableFieldConstants.MODE_NULLABLE,
-            description='INFO foo desc'),
-        bigquery.TableFieldSchema(
-            name='AB',
-            type=TableFieldConstants.TYPE_FLOAT,
-            mode=TableFieldConstants.MODE_NULLABLE,
-            description='INFO foo desc')
-    ]
-    merged_field_schemas = variant_to_bigquery._get_merged_field_schemas(
-        field_schemas_1, field_schemas_2)
-    expected_merged_field_schemas = [
-        bigquery.TableFieldSchema(
-            name='II',
-            type=TableFieldConstants.TYPE_INTEGER,
-            mode=TableFieldConstants.MODE_NULLABLE,
-            description='INFO foo desc'),
-        bigquery.TableFieldSchema(
-            name='IFR',
-            type=TableFieldConstants.TYPE_FLOAT,
-            mode=TableFieldConstants.MODE_REPEATED,
-            description='INFO foo desc'),
-        bigquery.TableFieldSchema(
-            name='AB',
-            type=TableFieldConstants.TYPE_FLOAT,
-            mode=TableFieldConstants.MODE_NULLABLE,
-            description='INFO foo desc')
-    ]
-    self.assertEqual(merged_field_schemas, expected_merged_field_schemas)
-
-  def test_merge_field_schemas_conflict_mode(self):
-    field_schemas_1 = [
-        bigquery.TableFieldSchema(
-            name='II',
-            type=TableFieldConstants.TYPE_INTEGER,
-            mode=TableFieldConstants.MODE_NULLABLE,
-            description='INFO foo desc')
-    ]
-    field_schemas_2 = [
-        bigquery.TableFieldSchema(
-            name='II',
-            type=TableFieldConstants.TYPE_INTEGER,
-            mode=TableFieldConstants.MODE_REPEATED,
-            description='INFO foo desc')
-    ]
-    self.assertRaises(ValueError, variant_to_bigquery._get_merged_field_schemas,
-                      field_schemas_1, field_schemas_2)
-
-  def test_merge_field_schemas_conflict_type(self):
-    field_schemas_1 = [
-        bigquery.TableFieldSchema(
-            name='II',
-            type=TableFieldConstants.TYPE_INTEGER,
-            mode=TableFieldConstants.MODE_NULLABLE,
-            description='INFO foo desc')
-    ]
-    field_schemas_2 = [
-        bigquery.TableFieldSchema(
-            name='II',
-            type=TableFieldConstants.TYPE_FLOAT,
-            mode=TableFieldConstants.MODE_NULLABLE,
-            description='INFO foo desc')
-    ]
-    self.assertRaises(ValueError, variant_to_bigquery._get_merged_field_schemas,
-                      field_schemas_1, field_schemas_2)
-
-  def test_merge_field_schemas_conflict_record_fields(self):
-    call_record_1 = bigquery.TableFieldSchema(
-        name=ColumnKeyConstants.CALLS,
-        type=TableFieldConstants.TYPE_RECORD,
-        mode=TableFieldConstants.MODE_REPEATED,
-        description='One record for each call.')
-    call_record_1.fields.append(bigquery.TableFieldSchema(
-        name='FB',
-        type=TableFieldConstants.TYPE_BOOLEAN,
-        mode=TableFieldConstants.MODE_NULLABLE,
-        description='FORMAT foo desc'))
-    field_schemas_1 = [call_record_1]
-
-    call_record_2 = bigquery.TableFieldSchema(
-        name=ColumnKeyConstants.CALLS,
-        type=TableFieldConstants.TYPE_RECORD,
-        mode=TableFieldConstants.MODE_REPEATED,
-        description='One record for each call.')
-    call_record_2.fields.append(bigquery.TableFieldSchema(
-        name='FB',
-        type=TableFieldConstants.TYPE_INTEGER,
-        mode=TableFieldConstants.MODE_NULLABLE,
-        description='FORMAT foo desc'))
-    field_schemas_2 = [call_record_2]
-    self.assertRaises(ValueError, variant_to_bigquery._get_merged_field_schemas,
-                      field_schemas_1, field_schemas_2)
-
-  def test_merge_field_schemas_same_record(self):
-    call_record_1 = bigquery.TableFieldSchema(
-        name=ColumnKeyConstants.CALLS,
-        type=TableFieldConstants.TYPE_RECORD,
-        mode=TableFieldConstants.MODE_REPEATED,
-        description='One record for each call.')
-    call_record_1.fields.append(bigquery.TableFieldSchema(
-        name='FB',
-        type=TableFieldConstants.TYPE_BOOLEAN,
-        mode=TableFieldConstants.MODE_NULLABLE,
-        description='FORMAT foo desc'))
-
-    field_schemas_1 = [call_record_1]
-    field_schemas_2 = [call_record_1]
-
-    expected_merged_field_schemas = [call_record_1]
-    self.assertEqual(
-        variant_to_bigquery._get_merged_field_schemas(field_schemas_1,
-                                                      field_schemas_2),
-        expected_merged_field_schemas)
-
-  def test_merge_field_schemas_merge_record_fields(self):
-    call_record_1 = bigquery.TableFieldSchema(
-        name=ColumnKeyConstants.CALLS,
-        type=TableFieldConstants.TYPE_RECORD,
-        mode=TableFieldConstants.MODE_REPEATED,
-        description='One record for each call.')
-    call_record_1.fields.append(bigquery.TableFieldSchema(
-        name='FB',
-        type=TableFieldConstants.TYPE_BOOLEAN,
-        mode=TableFieldConstants.MODE_NULLABLE,
-        description='FORMAT foo desc'))
-
-    field_schemas_1 = [call_record_1]
-
-    call_record_2 = bigquery.TableFieldSchema(
-        name=ColumnKeyConstants.CALLS,
-        type=TableFieldConstants.TYPE_RECORD,
-        mode=TableFieldConstants.MODE_REPEATED,
-        description='One record for each call.')
-    call_record_2.fields.append(bigquery.TableFieldSchema(
-        name='GQ',
-        type=TableFieldConstants.TYPE_INTEGER,
-        mode=TableFieldConstants.MODE_NULLABLE,
-        description='FORMAT foo desc'))
-    field_schemas_2 = [call_record_2]
-
-    call_record_3 = bigquery.TableFieldSchema(
-        name=ColumnKeyConstants.CALLS,
-        type=TableFieldConstants.TYPE_RECORD,
-        mode=TableFieldConstants.MODE_REPEATED,
-        description='One record for each call.')
-    call_record_3.fields.append(bigquery.TableFieldSchema(
-        name='FB',
-        type=TableFieldConstants.TYPE_BOOLEAN,
-        mode=TableFieldConstants.MODE_NULLABLE,
-        description='FORMAT foo desc'))
-    call_record_3.fields.append(bigquery.TableFieldSchema(
-        name='GQ',
-        type=TableFieldConstants.TYPE_INTEGER,
-        mode=TableFieldConstants.MODE_NULLABLE,
-        description='FORMAT foo desc'))
-
-    expected_merged_field_schemas = [call_record_3]
-    self.assertEqual(
-        variant_to_bigquery._get_merged_field_schemas(field_schemas_1,
-                                                      field_schemas_2),
-        expected_merged_field_schemas)
-
-  def test_merge_field_schemas_conflict_inner_record_fields(self):
-    record_1 = bigquery.TableFieldSchema(
-        name=ColumnKeyConstants.CALLS,
-        type=TableFieldConstants.TYPE_RECORD,
-        mode=TableFieldConstants.MODE_REPEATED,
-        description='One record for each call.')
-    inner_record_1 = bigquery.TableFieldSchema(
-        name='inner record',
-        type=TableFieldConstants.TYPE_RECORD,
-        mode=TableFieldConstants.MODE_REPEATED,
-        description='One record for each call.')
-    inner_record_1.fields.append(bigquery.TableFieldSchema(
-        name='FB',
-        type=TableFieldConstants.TYPE_RECORD,
-        mode=TableFieldConstants.MODE_REPEATED,
-        description='FORMAT foo desc'))
-    record_1.fields.append(inner_record_1)
-    field_schemas_1 = [record_1]
-
-    record_2 = bigquery.TableFieldSchema(
-        name=ColumnKeyConstants.CALLS,
-        type=TableFieldConstants.TYPE_RECORD,
-        mode=TableFieldConstants.MODE_REPEATED,
-        description='One record for each call.')
-    inner_record_2 = bigquery.TableFieldSchema(
-        name='inner record',
-        type=TableFieldConstants.TYPE_INTEGER,
-        mode=TableFieldConstants.MODE_REPEATED,
-        description='One record for each call.')
-    inner_record_2.fields.append(bigquery.TableFieldSchema(
-        name='FB',
-        type=TableFieldConstants.TYPE_RECORD,
-        mode=TableFieldConstants.MODE_REPEATED,
-        description='FORMAT foo desc'))
-    record_2.fields.append(inner_record_2)
-    field_schemas_2 = [record_2]
-    self.assertRaises(ValueError, variant_to_bigquery._get_merged_field_schemas,
-                      field_schemas_1, field_schemas_2)
-
-  def test_merge_field_schemas_merge_inner_record_fields(self):
-    record_1 = bigquery.TableFieldSchema(
-        name=ColumnKeyConstants.CALLS,
-        type=TableFieldConstants.TYPE_RECORD,
-        mode=TableFieldConstants.MODE_REPEATED,
-        description='One record for each call.')
-    inner_record_1 = bigquery.TableFieldSchema(
-        name='inner record',
-        type=TableFieldConstants.TYPE_RECORD,
-        mode=TableFieldConstants.MODE_REPEATED,
-        description='One record for each call.')
-    inner_record_1.fields.append(bigquery.TableFieldSchema(
-        name='FB',
-        type=TableFieldConstants.TYPE_RECORD,
-        mode=TableFieldConstants.MODE_REPEATED,
-        description='FORMAT foo desc'))
-    record_1.fields.append(inner_record_1)
-    field_schemas_1 = [record_1]
-
-    record_2 = bigquery.TableFieldSchema(
-        name=ColumnKeyConstants.CALLS,
-        type=TableFieldConstants.TYPE_RECORD,
-        mode=TableFieldConstants.MODE_REPEATED,
-        description='One record for each call.')
-    inner_record_2 = bigquery.TableFieldSchema(
-        name='inner record',
-        type=TableFieldConstants.TYPE_RECORD,
-        mode=TableFieldConstants.MODE_REPEATED,
-        description='One record for each call.')
-    inner_record_2.fields.append(bigquery.TableFieldSchema(
-        name='AB',
-        type=TableFieldConstants.TYPE_RECORD,
-        mode=TableFieldConstants.MODE_REPEATED,
-        description='FORMAT foo desc'))
-    record_2.fields.append(inner_record_2)
-    field_schemas_2 = [record_2]
-
-    merged_record = bigquery.TableFieldSchema(
-        name=ColumnKeyConstants.CALLS,
-        type=TableFieldConstants.TYPE_RECORD,
-        mode=TableFieldConstants.MODE_REPEATED,
-        description='One record for each call.')
-    merged_inner_record = bigquery.TableFieldSchema(
-        name='inner record',
-        type=TableFieldConstants.TYPE_RECORD,
-        mode=TableFieldConstants.MODE_REPEATED,
-        description='One record for each call.')
-    merged_inner_record.fields.append(bigquery.TableFieldSchema(
-        name='FB',
-        type=TableFieldConstants.TYPE_RECORD,
-        mode=TableFieldConstants.MODE_REPEATED,
-        description='FORMAT foo desc'))
-    merged_inner_record.fields.append(bigquery.TableFieldSchema(
-        name='AB',
-        type=TableFieldConstants.TYPE_RECORD,
-        mode=TableFieldConstants.MODE_REPEATED,
-        description='FORMAT foo desc'))
-    merged_record.fields.append(merged_inner_record)
-    expected_merged_field_schemas = [merged_record]
-    self.assertEqual(
-        variant_to_bigquery._get_merged_field_schemas(field_schemas_1,
-                                                      field_schemas_2),
-        expected_merged_field_schemas)
